requirements.txt
py-rrdtool>=0.2
pyzmq>=14
cliff>=1
